requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy==1.23.5
tensorflow-cpu==2.10.0